WIP - resizing image subtitles to fit video frame
It often seems that image subtitles are deliberately at a different resolution to the video file, which leads to them burning into the top left quarter of the frame. ffmpeg supports a way to resize the subtitles to fit the frame; this is a WIP change to support this in cine-encoder.
Due to the potential to have the video size also changing, this required a few other changes that need to be tested, hence the WIP note. This is currently just compile-tested pending the outcome.